requirements.txt
Flask==3.0.3
requests==2.32.3
python-telegram-bot==21.3
Werkzeug==3.0.3
